Pushed all the UI to the windows.
