+++ managers/game_manager.py
@@ -19,8 +19,8 @@ class GameManager(object):
         self.game_context = GameContext()
 
         self.console_manager = ConsoleManager()
-        self.scene_manager = SceneManager(self.console_manager,
-                                          game_context=self.game_context)
+        self.scene_manager = SceneManager(self.console_manager, game_context=self.game_context)
+        self.game_context.console_manager = self.console_manager
         self.load_game_data()
 
     def start(self):
