+++ scenes/__init__.py
@@ -1,5 +1,3 @@
-from .game import GameScene
-from .inventory import InventoryScene
-from .inventory_query import InventoryQueryScene
-from .main_menu import MainMenuScene
-from .character_creation import CharacterCreationScene
+from scenes.character_creation.scene import CharacterCreationScene
+from scenes.game.scene import GameScene
+from scenes.main_menu.scene import MainMenuScene
