+++ scenes/game/windows/__init__.py
@@ -1,0 +1,3 @@
+from scenes.game.windows.game_window import GameWindow
+from scenes.game.windows.inventory_window import InventoryWindow
+from scenes.game.windows.item_query_window import ItemQueryWindow
